[Refactor] photos: 옵저버 분리
